✨(admin) add admin views to manage referral urgencies
As referral urgency levels are supposed to be managed through
the admin, we need to add admin views for them. We're using django
parler's utility class coupled with a Tabular view for translations.

This way, the admin provides a quick and easy way to see all
translations and modify them on the spot.